COMMERCE-2769 - multiple refactors to comply to Mulesoft rules
